+++ daemon/daemon_windows.go
@@ -16,6 +16,7 @@ import (
 	"github.com/docker/docker/daemon/container"
 	"github.com/docker/docker/daemon/internal/libcontainerd/local"
 	"github.com/docker/docker/daemon/internal/libcontainerd/remote"
+	"github.com/docker/docker/daemon/internal/system"
 	"github.com/docker/docker/daemon/libnetwork"
 	nwconfig "github.com/docker/docker/daemon/libnetwork/config"
 	winlibnetwork "github.com/docker/docker/daemon/libnetwork/drivers/windows"
@@ -25,7 +26,6 @@ import (
 	"github.com/docker/docker/daemon/network"
 	"github.com/docker/docker/pkg/parsers/operatingsystem"
 	"github.com/docker/docker/pkg/sysinfo"
-	"github.com/docker/docker/pkg/system"
 	containertypes "github.com/moby/moby/api/types/container"
 	networktypes "github.com/moby/moby/api/types/network"
 	"github.com/moby/sys/user"
