+++ daemon/graphdriver/copy/copy.go
@@ -12,8 +12,8 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/docker/docker/daemon/internal/system"
 	"github.com/docker/docker/pkg/pools"
-	"github.com/docker/docker/pkg/system"
 	"github.com/moby/sys/userns"
 	"golang.org/x/sys/unix"
 )
