+++ daemon/internal/image/tarexport/os_path.go
@@ -14,7 +14,7 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/docker/docker/pkg/system"
+	"github.com/docker/docker/daemon/internal/system"
 )
 
 // mkdirAllWithChtimes is nearly an identical copy to the [os.MkdirAll] but
